requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
dataclasses-json==0.5.14
decorator==5.1.1
dnspython==2.4.2
elevenlabs==0.2.24
exceptiongroup==1.1.3
executing==1.2.0
firebase-admin==6.2.0
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
langchain==0.0.279
langsmith==0.0.32
loguru==0.7.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
pinecone-client==2.2.2
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.2
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==3.15.4
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.20
stack-data==0.6.2
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.66.1
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.7
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.16.2
